+++ docs/changes.md
@@ -2,6 +2,7 @@
 * 2019/08/14
   * 完善文档。
   * 增加宏TK\_XML\_MAX\_ATTRS用于控制XML属性的最大个数(缺省24)。
+  * 通过widget factory触发EVT\_WIDGET\_CREATED事件。
 
 * 2019/08/13
   * 模拟器支持单色屏。